+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Added PodcastId to the list of episode GQL arguments.
 - Fixed a bug on generating RSS where there is a field for podcast live items but that field is not on podcast field layout.
 
 ## 0.20.0 - 2024-04-24